Look a Channel peer up by its ID
