+++ lib/java/os-environment/src/main/java/org/enso/os/environment/jni/TestMain.java
@@ -1,10 +1,7 @@
 package org.enso.os.environment.jni;
 
-import java.lang.foreign.FunctionDescriptor;
-import java.lang.foreign.Linker;
-import java.lang.foreign.MemorySegment;
-import java.lang.foreign.ValueLayout;
-import java.lang.invoke.MethodHandle;
+import java.io.File;
+import java.io.FileWriter;
 import java.math.BigInteger;
 import java.util.HashMap;
 import java.util.Map;
@@ -15,17 +12,12 @@ final class TestMain {
 
   private TestMain() {}
 
-  static void main(String... args) throws Throwable {
-    var jvmIsolate = Long.parseLong(args[0]);
-    var fnCallbackAddress = MemorySegment.ofAddress(Long.parseLong(args[1]));
-    var fnDescriptor =
-        FunctionDescriptor.of(
-            ValueLayout.JAVA_LONG, ValueLayout.ADDRESS, ValueLayout.ADDRESS, ValueLayout.JAVA_LONG);
-    var fnHandle = Linker.nativeLinker().downcallHandle(fnCallbackAddress, fnDescriptor);
-    var n = Integer.parseInt(args[2]);
-
-    var res = factorial(n).toString();
-    reportResultToSvmIsolate(jvmIsolate, fnHandle, n, res);
+  public static void main(String... args) throws Exception {
+    var out = new File(args[0]);
+    var n = Integer.parseInt(args[1]);
+    try (java.io.FileWriter os = new FileWriter(out)) {
+      os.write(factorial(n).toString());
+    }
   }
 
   static BigInteger factorial(long n) {
@@ -37,24 +29,6 @@ final class TestMain {
       }
     }
     return acc;
-  }
-
-  private static void reportResultToSvmIsolate(
-      long jvmIsolate, MethodHandle fn, long key, String value) throws Throwable {
-    ReportResult rr = new ReportResult(key, value);
-    JVM.executeImpl(
-        JVMPeer.POOL,
-        rr,
-        (seg) -> {
-          Object res = -1L;
-          try {
-            var isoRef = MemorySegment.ofAddress(jvmIsolate);
-            res = fn.invoke(isoRef, seg, seg.byteSize());
-          } catch (Throwable ex) {
-            ex.printStackTrace();
-          }
-          return (long) res;
-        });
   }
 
   @Persistable(id = 430607)
